optimize and adapter oracle
