+++ rm-datasource/src/main/java/io/seata/rm/datasource/sql/druid/BaseRecognizer.java
@@ -15,7 +15,13 @@
  */
 package io.seata.rm.datasource.sql.druid;
 
+import com.alibaba.druid.sql.ast.expr.SQLVariantRefExpr;
+import com.alibaba.druid.sql.dialect.mysql.visitor.MySqlOutputVisitor;
+import io.seata.rm.datasource.ParametersHolder;
 import io.seata.rm.datasource.sql.SQLRecognizer;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * The type Base recognizer.
@@ -53,5 +59,26 @@ public abstract class BaseRecognizer implements SQLRecognizer {
     @Override
     public String getOriginalSQL() {
         return originalSQL;
+    }
+
+    public MySqlOutputVisitor createMySqlOutputVisitor(final ParametersHolder parametersHolder, final ArrayList<List<Object>> paramAppenders, StringBuffer sb) {
+        MySqlOutputVisitor visitor = new MySqlOutputVisitor(sb) {
+
+            @Override
+            public boolean visit(SQLVariantRefExpr x) {
+                if ("?".equals(x.getName())) {
+                    ArrayList<Object> paramAppender = parametersHolder.getParameters()[x.getIndex()];
+                    if (paramAppenders.size() == 0) {
+                        paramAppender.stream().forEach(t -> paramAppenders.add(new ArrayList<>()));
+                    }
+                    for (int i = 0; i < paramAppender.size(); i++) {
+                        paramAppenders.get(i).add(paramAppender.get(i));
+                    }
+
+                }
+                return super.visit(x);
+            }
+        };
+        return visitor;
     }
 }
