+++ rm-datasource/src/main/java/io/seata/rm/datasource/sql/druid/MySQLSelectForUpdateRecognizer.java
@@ -52,7 +52,7 @@ public class MySQLSelectForUpdateRecognizer extends BaseRecognizer implements SQ
      */
     public MySQLSelectForUpdateRecognizer(String originalSQL, SQLStatement ast) {
         super(originalSQL);
-        this.ast = (SQLSelectStatement)ast;
+        this.ast = (SQLSelectStatement) ast;
     }
 
     @Override
@@ -68,25 +68,9 @@ public class MySQLSelectForUpdateRecognizer extends BaseRecognizer implements SQ
             return "";
         }
         StringBuffer sb = new StringBuffer();
-        MySqlOutputVisitor visitor = new MySqlOutputVisitor(sb) {
-
-            @Override
-            public boolean visit(SQLVariantRefExpr x) {
-                if ("?".equals(x.getName())) {
-                    ArrayList<Object> paramAppender = parametersHolder.getParameters()[x.getIndex()];
-                    if (paramAppenders.size() == 0) {
-                        paramAppender.stream().forEach(t -> paramAppenders.add(new ArrayList<Object>()));
-                    }
-                    for (int i = 0; i < paramAppender.size(); i++) {
-                        paramAppenders.get(i).add(paramAppender.get(i));
-                    }
-
-                }
-                return super.visit(x);
-            }
-        };
+        MySqlOutputVisitor visitor = super.createMySqlOutputVisitor(parametersHolder, paramAppenders, sb);
         if (where instanceof SQLBinaryOpExpr) {
-            visitor.visit((SQLBinaryOpExpr)where);
+            visitor.visit((SQLBinaryOpExpr) where);
         } else if (where instanceof SQLInListExpr) {
             visitor.visit((SQLInListExpr) where);
         } else if (where instanceof SQLBetweenExpr) {
@@ -106,7 +90,7 @@ public class MySQLSelectForUpdateRecognizer extends BaseRecognizer implements SQ
         }
         StringBuffer sb = new StringBuffer();
         MySqlOutputVisitor visitor = new MySqlOutputVisitor(sb);
-        visitor.visit((SQLBinaryOpExpr)where);
+        visitor.visit((SQLBinaryOpExpr) where);
         return sb.toString();
     }
 
@@ -142,7 +126,7 @@ public class MySQLSelectForUpdateRecognizer extends BaseRecognizer implements SQ
                 return false;
             }
         };
-        visitor.visit((SQLExprTableSource)tableSource);
+        visitor.visit((SQLExprTableSource) tableSource);
         return sb.toString();
     }
